move all build functions except func to dialect
